% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{add_assessment_derived_cols}
\alias{add_assessment_derived_cols}
\title{Add derived columns to assessment data}
\usage{
add_assessment_derived_cols(df, end_year)
}
\arguments{
\item{df}{Data frame with standardized columns}

\item{end_year}{School year end}
}
\value{
Data frame with derived columns
}
\description{
Adds aggregation flags and other derived columns.
}
\keyword{internal}
