% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{build_directory_url}
\alias{build_directory_url}
\title{Build OSDE directory file URL}
\usage{
build_directory_url(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"District" or "Site"}
}
\value{
URL string
}
\description{
Constructs the URL for downloading directory data from OSDE.
Handles different file locations based on year and type.
}
\keyword{internal}
