% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_osde_url}
\alias{build_osde_url}
\title{Build OSDE enrollment file URL}
\usage{
build_osde_url(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"District" or "Site"}
}
\value{
URL string
}
\description{
Constructs the URL for downloading enrollment data from OSDE.
Handles different file naming patterns across eras.
}
\keyword{internal}
