% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{build_osde_url_alt}
\alias{build_osde_url_alt}
\title{Build alternate OSDE enrollment file URL}
\usage{
build_osde_url_alt(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"District" or "Site"}
}
\value{
URL string or NULL if no alternate
}
\description{
Constructs an alternate URL pattern for OSDE data.
}
\keyword{internal}
