% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{create_empty_directory_df}
\alias{create_empty_directory_df}
\title{Create empty directory data frame}
\usage{
create_empty_directory_df(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with correct structure
}
\description{
Create empty directory data frame
}
\keyword{internal}
