% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_by_grade}
\alias{download_assessment_by_grade}
\title{Download assessment data by grade}
\usage{
download_assessment_by_grade(end_year)
}
\arguments{
\item{end_year}{School year end (2017 or 2018)}
}
\value{
Combined data frame with all grades
}
\description{
Downloads grade-level assessment files for 2017-2018 and combines them.
}
\keyword{internal}
