% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_file}
\alias{download_assessment_file}
\title{Download a single assessment file}
\usage{
download_assessment_file(url, end_year)
}
\arguments{
\item{url}{URL to download}

\item{end_year}{School year end (for error messages)}
}
\value{
Data frame with assessment data
}
\description{
Download a single assessment file
}
\keyword{internal}
