% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_oklaschools_data}
\alias{download_oklaschools_data}
\title{Download OklaSchools.com report card data}
\usage{
download_oklaschools_data(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with report card data
}
\description{
Downloads enrollment and demographics data from OklaSchools.com Data Matrix.
This is used as a supplementary data source for demographics and grade levels.
}
\keyword{internal}
