% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{download_osde_directory}
\alias{download_osde_directory}
\title{Download OSDE directory Excel file}
\usage{
download_osde_directory(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"District" or "Site"}
}
\value{
Data frame with directory data
}
\description{
Downloads district or site directory data from OSDE.
}
\keyword{internal}
