% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_osde_enrollment}
\alias{download_osde_enrollment}
\title{Download OSDE enrollment Excel file}
\usage{
download_osde_enrollment(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"District" or "Site"}
}
\value{
Data frame with enrollment data
}
\description{
Downloads district or site enrollment data from OSDE.
}
\keyword{internal}
