% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{get_assessment_cache_path}
\alias{get_assessment_cache_path}
\title{Get assessment cache path}
\usage{
get_assessment_cache_path(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Cache type ("assessment_tidy" or "assessment_wide")}
}
\value{
Full path to cache file
}
\description{
Get assessment cache path
}
\keyword{internal}
