% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Oklahoma enrollment data}
\usage{
get_available_years()
}
\value{
Integer vector of available years
}
\description{
Returns a vector of school year ends for which enrollment data is available.
Data is available from 2016 (FY15-16) through 2025 (current).
}
\examples{
get_available_years()
}
