% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assessment}
\alias{get_raw_assessment}
\title{Download raw assessment data from OSDE}
\usage{
get_raw_assessment(end_year)
}
\arguments{
\item{end_year}{School year end (2017-2019, 2022-2025; no 2020-2021 due to COVID)}
}
\value{
Data frame with raw assessment data
}
\description{
Downloads OSTP assessment data from the Oklahoma State Department of
Education's public data files.
}
\keyword{internal}
