% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw directory data from OSDE}
\usage{
get_raw_directory(end_year)
}
\arguments{
\item{end_year}{School year end (2024-25 = 2025)}
}
\value{
List with district and site data frames
}
\description{
Downloads district and site directory data from OSDE's public data files.
}
\keyword{internal}
