% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_district_directory}
\alias{process_district_directory}
\title{Process district-level directory data}
\usage{
process_district_directory(df, end_year)
}
\arguments{
\item{df}{Raw district directory data frame}

\item{end_year}{School year end}
}
\value{
Processed district data frame
}
\description{
Process district-level directory data
}
\keyword{internal}
