% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_site_directory}
\alias{process_site_directory}
\title{Process site-level directory data}
\usage{
process_site_directory(df, end_year)
}
\arguments{
\item{df}{Raw site directory data frame}

\item{end_year}{School year end}
}
\value{
Processed site data frame
}
\description{
Process site-level directory data
}
\keyword{internal}
