% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_site_enr}
\alias{process_site_enr}
\title{Process site-level enrollment data}
\usage{
process_site_enr(df, end_year)
}
\arguments{
\item{df}{Raw site data frame}

\item{end_year}{School year end}
}
\value{
Processed site data frame
}
\description{
Sites in Oklahoma are equivalent to campuses/schools in other states.
}
\keyword{internal}
