% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_assessment.R
\name{read_assessment_cache}
\alias{read_assessment_cache}
\title{Read assessment data from cache}
\usage{
read_assessment_cache(end_year, type)
}
\arguments{
\item{end_year}{School year end}

\item{type}{Cache type}
}
\value{
Cached data frame
}
\description{
Read assessment data from cache
}
\keyword{internal}
