% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years for Oklahoma enrollment data}
\usage{
get_available_years()
}
\value{
A list with components:
\describe{
\item{min_year}{Earliest available school year end (2016)}
\item{max_year}{Latest available school year end (2024)}
\item{description}{Human-readable description of the data availability}
}
}
\description{
Returns metadata about the range of school years for which enrollment data
is available from the Oklahoma State Department of Education.
}
\examples{
years <- get_available_years()
years$min_year
years$max_year
years$description
}
